requirements.txt
discord.py==2.3.2
mcstatus==11.0.0